# api_doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "Eazika API - A comprehensive e-commerce platform connecting customers, shopkeepers, and delivery partners. This API provides authentication, shop management, product catalog, cart operations, order management, and delivery services.",
    "version": "2.0.0",
    "title": "Eazika API",
    "termsOfService": "https://eazika.com/terms/",
    "contact": {
      "email": "contact@eazika.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api/v2",
...
  "schemes": ["http", "https"],
  "paths": {
    "/users/register": {
      "post": {
        "tags": ["user"],
        "summary": "Send registration OTP",
        "description": "Initiates the user registration process by sending an OTP to the provided phone number",
        "operationId": "sendRegistrationOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Registration details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegistrationOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "schema": {
              "$ref": "#/definitions/OtpResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/register/verify": {
      "post": {
        "tags": ["user"],
        "summary": "Verify registration OTP",
        "description": "Verifies the OTP sent during registration and creates a new user account",
        "operationId": "verifyRegistrationOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "OTP verification details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP verified successfully, user created",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          },
          "400": {
            "description": "Invalid OTP or request"
          },
          "500": {
            "description": "User creation failed"
          }
        }
      }
    },
    "/users/register/resend": {
      "post": {
        "tags": ["user"],
        "summary": "Resend registration OTP",
        "description": "Resends the registration OTP to the provided phone number",
        "operationId": "resendRegistrationOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Registration details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegistrationOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP resent successfully",
            "schema": {
              "$ref": "#/definitions/OtpResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["user"],
        "summary": "Send login OTP",
        "description": "Initiates the login process by sending an OTP to the registered phone number",
        "operationId": "sendLoginOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login phone number",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "schema": {
              "$ref": "#/definitions/OtpResponse"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/login/verify": {
      "post": {
        "tags": ["user"],
        "summary": "Verify login OTP",
        "description": "Verifies the login OTP and issues authentication tokens",
        "operationId": "verifyLoginOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "OTP verification details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "$ref": "#/definitions/AuthResponse"
            }
          },
          "400": {
            "description": "Invalid OTP"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/login/resend": {
      "post": {
        "tags": ["user"],
        "summary": "Resend login OTP",
        "description": "Resends the login OTP to the provided phone number",
        "operationId": "resendLoginOtp",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login phone number",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginOtpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP resent successfully",
            "schema": {
              "$ref": "#/definitions/OtpResponse"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/refresh": {
      "post": {
        "tags": ["user"],
        "summary": "Refresh access token",
        "description": "Generates new access and refresh tokens using a valid refresh token",
        "operationId": "refreshToken",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Refresh token (optional if sent via cookie)",
            "required": false,
            "schema": {
              "$ref": "#/definitions/RefreshTokenRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "schema": {
              "$ref": "#/definitions/TokenResponse"
            }
          },
          "401": {
            "description": "Invalid or missing refresh token"
          }
        }
      }
    },
...
    "/users/user/me": {
      "get": {
        "tags": ["user"],
        "summary": "Get current user profile",
        "description": "Retrieves the profile information of the currently authenticated user",
        "operationId": "getCurrentUser",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": ["user"],
        "summary": "Update current user profile",
        "description": "Updates the profile information of the currently authenticated user",
        "operationId": "updateCurrentUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Updated user details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token in the format: Bearer <token>"
    }
  },
  "definitions": {
...
    "RegistrationOtpRequest": {
      "type": "object",
      "required": ["name", "phone"],
      "properties": {
        "name": {
          "type": "string",
          "minLength": 2,
          "maxLength": 100,
          "example": "John Doe"
        },
        "phone": {
          "type": "string",
          "pattern": "^\\d{10}$",
          "example": "9876543210"
        },
        "deviceInfo": {
          "type": "string",
          "example": "iPhone 13 Pro"
        }
      }
    },
    "LoginOtpRequest": {
      "type": "object",
      "required": ["phone"],
      "properties": {
        "phone": {
          "type": "string",
          "pattern": "^\\d{10}$",
          "example": "9876543210"
        }
      }
    },
    "VerifyOtpRequest": {
      "type": "object",
      "required": ["phone", "requestId", "otp"],
      "properties": {
        "phone": {
          "type": "string",
          "pattern": "^\\d{10}$",
          "example": "9876543210"
        },
        "requestId": {
          "type": "string",
          "example": "req_1234567890"
        },
        "otp": {
          "type": "string",
          "minLength": 4,
          "maxLength": 4,
          "example": "1234"
        },
        "deviceInfo": {
          "type": "string",
          "example": "iPhone 13 Pro"
        }
      }
    },
    "OtpResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "integer",
          "format": "int32",
          "example": 200
        },
        "message": {
          "type": "string",
          "example": "OTP sent successfully"
        },
        "data": {
          "type": "object",
          "properties": {
            "requestId": {
              "type": "string",
              "example": "req_1234567890"
            }
          }
        }
      }
    },
    "AuthResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "integer",
          "format": "int32",
          "example": 200
        },
        "message": {
          "type": "string",
          "example": "Authentication successful"
        },
        "data": {
          "type": "object",
          "properties": {
            "accessToken": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            },
            "expiresIn": {
              "type": "string",
              "example": "7d"
            },
            "user": {
              "$ref": "#/definitions/UserBasic"
            }
          }
        }
      }
    },
    "TokenResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "integer",
          "format": "int32",
          "example": 200
        },
        "message": {
          "type": "string",
          "example": "Token refreshed successfully"
        },
        "data": {
          "type": "object",
          "properties": {
            "accessToken": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            },
            "expiresIn": {
              "type": "string",
              "example": "7d"
            }
          }
        }
      }
    },
    "RefreshTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "phone": {
          "type": "string",
          "example": "9876543210"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "image": {
          "type": "string",
          "example": "https://example.com/profile.jpg"
        },
        "role": {
          "type": "string",
          "enum": ["user", "shopkeeper", "delivery_boy", "admin"],
          "example": "user"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "isPhoneVerified": {
          "type": "boolean",
          "example": true
        },
        "isEmailVerified": {
          "type": "boolean",
          "example": false
        },
        "defaultAddressId": {
          "type": "integer",
          "format": "int32",
          "example": 1
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "UserBasic": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1
        },
        "phone": {
          "type": "string",
          "example": "9876543210"
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "image": {
          "type": "string",
          "example": "https://example.com/profile.jpg"
        }
      }
    },
...
  },
...
}